requirements.txt
boto3==1.13.1
celery==4.4.6
Django==3.0.5
django-admin-honeypot==1.1.0
django-apptemplates==1.5
django-debug-toolbar==2.2
django-dotenv==1.4.2
django-environ==0.4.5
django-mptt==0.11.0
django-redis-cache==2.1.1
django-storages==1.9.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
whitenoise==5.0.1
